dynamic_agent_service/pyproject.toml
[project]
name = "dynamic-agent-service"
version = "0.1.0"
description = "Backend service that orchestrates the agent, manages conversations, and coordinates tool execution"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pyjwt>=2.8",
    "pydantic>=2.0",
    "colorama>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "util"]